refactor(client/map): 커스텀오버레이 호버 이벤트 시의 로직 모듈화
함수로 빼서 맵유틸 모듈에 추가
